make test utilites private
