+++ tests/_asserts.py
@@ -3,7 +3,7 @@ from urllib.request import Request, urlopen
 
 from torchvision.datasets.utils import calculate_md5
 
-from .utils import get_tmp_dir
+from ._utils import get_tmp_dir
 
 __all__ = ["assert_image_is_downloadable", "assert_image_downloads_correctly"]
